Remove copy and paste related to Async ActionNode
Remove copy and paste related to TickEngine and ActionNode

Instead ofcopy and pasting in any ControlNode the same specific code to
be used only in ActionNodes, put that code only once in the ActionNode
itself.

The only difference in the API is that the Actions implemented by the
user must implement the virtual funtion asyncTick(), instead of tick().

TickEngine was moved from TreeNode to ActionNode.